+++ CHANGELOG.rst
@@ -37,6 +37,7 @@ Added
 - Subscribed to ``kytos/mef_eline.link_available_tags`` events
 - Added ``_load_intf_available_tags`` to try to load and set available_Tags
 - Added ``_get_links_dict_with_tags`` to also have interface available_tags
+- Hooked ``_load_intf_available_tags`` to be called for interface_created
 
 [3.9.0] - 2021-12-22
 ********************